Bumped to version 2.1 and add checks while building RPMs

+++ fasthadd.spec
@@ -1,8 +1,9 @@
-%global commit d0b06280410956a3f98d42cfb40fc9ae4a717ac5
+%global commit a6cfbd2fad5011b6fdb08902efa3595cdc331e7e
 %global shortcommit %(c=%{commit}; echo ${c:0:7})
+%define files_for_build fastHadd.cc run_fastHadd_tests.sh test_fastHaddMerge.py fastParallelHadd.py
 
 Name:           fasthadd
-Version:        1.0
+Version:        2.1
 Release:        1%{?dist}
 Summary:        A program to add ProtocolBuffer-formatted ROOT files in a quick way
 
@@ -12,8 +13,8 @@ URL:            https://github.com/cms-sw/cmssw
 Source0:        https://github.com/rovere/cmssw/archive/%{commit}/%{name}-%{version}-%{shortcommit}.tar.gz
 BuildRoot:      %{_topdir}/BUILDROOT/%{name}-%{version}-%{release}
 
-BuildRequires:  root, protobuf-devel >= 2.4.1, protobuf-compiler >= 2.4.1
-Requires:       root, protobuf >= 2.4.1
+BuildRequires:  root, root-physics, root-graf3d, root-tree-player, protobuf-devel >= 2.4.1, protobuf-compiler >= 2.4.1
+Requires:       root, root-tree-player, protobuf >= 2.4.1
 
 %description
 A program to add ProtocolBuffer-formatted ROOT files in a quick way
@@ -25,11 +26,11 @@ A program to add ProtocolBuffer-formatted ROOT files in a quick way
 %build
 mkdir %{name}
 cd %{name}
-cp %{_builddir}/cmssw-%{commit}/DQMServices/Components/test/fastHadd.cc .
+for f in %{files_for_build}; do cp %{_builddir}/cmssw-%{commit}/DQMServices/Components/test/${f} .; done
 sed -i -e s#DQMServices/Core/src/ROOTFilePB.pb.h#ROOTFilePB.pb.h# fastHadd.cc
 cp %{_builddir}/cmssw-%{commit}/DQMServices/Core/src/ROOTFilePB.proto .
 protoc -I ./ --cpp_out=./ ROOTFilePB.proto
-g++ -O2 -o fasthadd ROOTFilePB.pb.cc fastHadd.cc `pkg-config --libs protobuf` `root-config --cflags --libs | sed -e s/-lGraf3d// | sed -e s/-lPhysics//`
+g++ -O2 -o fastHadd ROOTFilePB.pb.cc fastHadd.cc `pkg-config --libs protobuf` `root-config --cflags --libs`
 
 
 #make %{?_smp_mflags}
@@ -38,8 +39,20 @@ g++ -O2 -o fasthadd ROOTFilePB.pb.cc fastHadd.cc `pkg-config --libs protobuf` `r
 %install
 rm -rf %{buildroot}
 mkdir -p %{buildroot}%{_bindir}/
-cp -p %{name}/fasthadd %{buildroot}%{_bindir}/
+cp -p %{name}/fastHadd %{buildroot}%{_bindir}/
 
+%check
+mkdir -p test
+pushd test
+cp ../%{name}/fastHadd .
+for f in %{files_for_build}; do cp %{_builddir}/cmssw-%{commit}/DQMServices/Components/test/${f} .; done
+export PATH=./:${PATH}
+. ./run_fastHadd_tests.sh
+if [ $? != 0 ]; then
+  exit $?
+fi
+popd
+rm -fr test
 
 %clean
 rm -rf $RPM_BUILD_ROOT
@@ -48,7 +61,7 @@ rm -rf $RPM_BUILD_ROOT
 %files
 %defattr(-,root,root,-)
 %doc
-%{_bindir}/*
+%{_bindir}/fastHadd
 
 
 %changelog
